+++ src-bin/console.ts
@@ -1,10 +1,4 @@
-import { relative } from "path";
-
-export function log(pattern: string | Error, ...args: any[]): void {
+export function log(msg: any): void {
   // tslint:disable-next-line: no-console
-  console.log(pattern, ...args);
-}
-
-export function logFilename(filename: string, cwd: string, pattern: string) {
-  log(pattern, relative(cwd, filename));
+  console.log(msg);
 }