+++ src-bin/types.ts
@@ -1,5 +1,3 @@
-import { EventEmitter } from "events";
-import { RollupOptions } from "rollup";
 
 export interface BundlibOptions {
   dev?: boolean;
@@ -7,24 +5,10 @@ export interface BundlibOptions {
   silent?: boolean;
 }
 
-export type EventBuilding = "BUILDING";
-export type EventBuilt = "BUILT";
-
-export type EventWriting = "WRITING";
-export type EventWritten = "WRITTEN";
-
-export type EventError = "ERROR";
-
-export interface BuildEventEmitter extends EventEmitter {
-
-  on(event: EventWriting | EventWritten, listener: (filename: string) => void): this;
-  on(event: EventError, listener: (error: Error) => void): this;
-  on(event: EventBuilding | EventBuilt, listener: () => void): this;
-
-  emit(event: EventWriting | EventWritten, filename: string): boolean;
-  emit(event: EventError, error: Error): boolean;
-  emit(event: EventBuilding | EventBuilt): boolean;
-
+export interface BuildCallbackObject {
+  start?: () => void;
+  end?: () => void;
+  buildStart?: (filename: string) => void;
+  buildEnd?: (writeInfo: { filename: string, duration: number, size: number }) => void;
+  error?: (error: Error) => void;
 }
-
-export type BuldFunction = (configs: RollupOptions[]) => BuildEventEmitter;